+++ Demo/src/main/java/com/braintreepayments/demo/PayPalMessagingFragment.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.LinearLayout
 import com.braintreepayments.api.ExperimentalBetaApi
 import com.braintreepayments.api.PayPalMessagingColor
 import com.braintreepayments.api.PayPalMessagingListener
@@ -11,6 +12,7 @@ import com.braintreepayments.api.PayPalMessagingLogoType
 import com.braintreepayments.api.PayPalMessagingOfferType
 import com.braintreepayments.api.PayPalMessagingRequest
 import com.braintreepayments.api.PayPalMessagingTextAlignment
+import com.braintreepayments.api.PayPalMessagingView
 
 @OptIn(ExperimentalBetaApi::class)
 class PayPalMessagingFragment : BaseFragment(), PayPalMessagingListener {
@@ -35,17 +37,16 @@ class PayPalMessagingFragment : BaseFragment(), PayPalMessagingListener {
             color = PayPalMessagingColor.BLACK
         )
 
-        // TODO: Refactor BraintreeClient out of PayPalMessagingView
-//        val payPalMessagingView = PayPalMessagingView(braintreeClient, requireActivity())
-//        payPalMessagingView.setListener(this)
-//        payPalMessagingView.start(payPalMessagingRequest)
-//        payPalMessagingView.layoutParams = ViewGroup.LayoutParams(
-//            ViewGroup.LayoutParams.MATCH_PARENT,
-//            ViewGroup.LayoutParams.MATCH_PARENT
-//        )
-//
-//        val messagingView: LinearLayout = view.findViewById(R.id.content)
-//        messagingView.addView(payPalMessagingView)
+        val payPalMessagingView = PayPalMessagingView(requireContext(), authStringArg)
+        payPalMessagingView.setListener(this)
+        payPalMessagingView.start(payPalMessagingRequest)
+        payPalMessagingView.layoutParams = ViewGroup.LayoutParams(
+            ViewGroup.LayoutParams.MATCH_PARENT,
+            ViewGroup.LayoutParams.MATCH_PARENT
+        )
+
+        val messagingView: LinearLayout = view.findViewById(R.id.content)
+        messagingView.addView(payPalMessagingView)
     }
 
     override fun onPayPalMessagingClick() {