+++ ShopperInsights/src/main/java/com/braintreepayments/api/ShopperInsightsClient.kt
@@ -1,5 +1,6 @@
 package com.braintreepayments.api
 
+import android.content.Context
 import androidx.annotation.VisibleForTesting
 import com.braintreepayments.api.ShopperInsightsAnalytics.GET_RECOMMENDED_PAYMENTS_FAILED
 import com.braintreepayments.api.ShopperInsightsAnalytics.GET_RECOMMENDED_PAYMENTS_STARTED
@@ -24,7 +25,12 @@ class ShopperInsightsClient @VisibleForTesting internal constructor(
     private val api: ShopperInsightsApi,
     private val braintreeClient: BraintreeClient
 ) {
-    constructor(braintreeClient: BraintreeClient) : this(
+    /**
+     * @param context: an Android context
+     * @param authorization: a Tokenization Key or Client Token used to authenticate
+     */
+    constructor(context: Context, authorization: String) : this (BraintreeClient(context, authorization))
+    @VisibleForTesting internal constructor(braintreeClient: BraintreeClient) : this(
         ShopperInsightsApi(EligiblePaymentsApi(braintreeClient)),
         braintreeClient,
     )
@@ -32,7 +38,6 @@ class ShopperInsightsClient @VisibleForTesting internal constructor(
     /**
      * Retrieves recommended payment methods based on the provided shopper insights request.
      *
-     * @param context Android context
      * @param request The [ShopperInsightsRequest] containing information about the shopper.
      * @return A [ShopperInsightsResult] object indicating the recommended payment methods.
      * Note: This feature is in beta. Its public API may change or be removed in future releases